auto/requirements.txt
pytest==8.1.1
selenium==4.20.0
webdriver-manager==4.0.1
allure-pytest==2.13.5
allure-python-commons==2.13.5
requests==2.31.0    
bs4==0.0.2
flake8==7.1.0
